chore: 버튼 공통 디자인 폴더 common으로 이동
