+++ src/utils/common/getButtonProps.ts
@@ -1,0 +1,10 @@
+import { ReactNode } from 'react';
+
+export const getProps = (
+  stringProps: string | undefined,
+  nodeProps: ReactNode | undefined,
+): string | ReactNode | null => {
+  if (stringProps) return stringProps;
+  if (nodeProps) return nodeProps;
+  return null;
+};
